Changes done in Instagram_video_downloader.py
Please follow the formatting guidelines.
Make sure strings are single quotes
Python indentation is two spaces
'Please enter the video ID!'
You only need one line between the parts of the code
Just return instead of sys.exit()
